[refactor] khushal87 Overlay, Pricing
