+++ src/Overlay/Overlay.tsx
@@ -9,7 +9,6 @@ import {
   ViewStyle,
   StyleProp,
 } from 'react-native';
-import { withTheme } from '../config';
 import { RneFunctionComponent } from '../helpers';
 
 export type OverlayProps = ModalProps & {
@@ -21,12 +20,12 @@ export type OverlayProps = ModalProps & {
   ModalComponent?: typeof React.Component;
 };
 
-const Overlay: RneFunctionComponent<OverlayProps> = ({
+export const Overlay: RneFunctionComponent<OverlayProps> = ({
   children,
   backdropStyle,
   overlayStyle,
-  onBackdropPress,
-  fullScreen,
+  onBackdropPress = () => null,
+  fullScreen = false,
   ModalComponent = Modal,
   isVisible,
   ...rest
@@ -60,12 +59,6 @@ const Overlay: RneFunctionComponent<OverlayProps> = ({
     </View>
   </ModalComponent>
 );
-
-Overlay.defaultProps = {
-  fullScreen: false,
-  onBackdropPress: () => null,
-  ModalComponent: Modal,
-};
 
 const styles = StyleSheet.create({
   backdrop: {
@@ -104,5 +97,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export { Overlay };
-export default withTheme(Overlay, 'Overlay');
+Overlay.displayName = 'Overlay';
