+++ src/Overlay/index.tsx
@@ -1,0 +1,6 @@
+import { withTheme } from '../config';
+import { Overlay, OverlayProps } from './Overlay';
+
+export { Overlay };
+export type { OverlayProps };
+export default withTheme(Overlay, 'Overlay');
