+++ src/Pricing/__tests__/PricingCard.tsx
@@ -4,7 +4,7 @@ import toJson from 'enzyme-to-json';
 import { create } from 'react-test-renderer';
 import theme from '../../config/theme';
 import { ThemeProvider } from '../../config';
-import ThemedPricingCard, { PricingCard, PricingButton } from '../PricingCard';
+import ThemedPricingCard, { PricingCard, PricingButton } from '../index';
 import Icon from '../../icons/Icon';
 import Button from '../../Button';
