+++ src/config/theme.tsx
@@ -14,8 +14,8 @@ import { IconProps } from '../icons/Icon';
 import { ImageProps } from '../image/Image';
 import { InputProps } from '../input/Input';
 import { ListItemProps } from '../list/ListItemBase';
-import { OverlayProps } from '../overlay/Overlay';
-import { PricingCardProps } from '../pricing/PricingCard';
+import { OverlayProps } from '../Overlay';
+import { PricingCardProps } from '../Pricing';
 import { SearchBarProps } from '../searchbar/SearchBar';
 import { SliderProps } from '../Slider';
 import { SocialIconProps } from '../SocialIcon';
